Label spaces and fetch by label
There are some default spaces created by the CATs setup. These spaces
get a randomised GUID which is the default sorting field which means
sometimes the first 100 results for /v3/spaces includes a non-perf space
and then the GetGUIDs functions filters this space out. This causes some
assertions that 100 results were returned to fail (in the
security_groups test)